requirements.txt
tensorflow-gpu~=2.2.0
pandas~=0.25.1
sklearn~=0.0
scikit-learn~=0.21.2
imgaug~=0.3.0
numpy~=1.17.1
Keras~=2.3.1
tqdm~=4.32.2
opencv-python~=4.1.1.26
matplotlib~=3.1.0
natsort~=6.2.0
statsmodels~=0.10.0
scikit-image~=0.16.2
hiplot~=0.1.10
kerastuner~=1.0.1